+++ FileReader.h
@@ -10,7 +10,7 @@ protected:
 public:
 	static bool Exist(_LPCTSTR fname);
 	size_t Read(void * data, long count);	
-	bool GetLine(_LPTSTR data, long count);
+	bool GetLine(_LPTSTR data, int count);
 	static CFileReader * Open(_LPCTSTR fname, long mode = FILE_READ | FILE_BIN);
 	bool Eof(void);
 };
